requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
django-cors-headers>=3.2.1,<3.3.0
psycopg2>=2.8.4,<2.9.0
flake8>=3.7.9,<3.8.0
isort>=5.10.0,<5.20.0
channels==3.0.5
channels-redis==4.0.0
daphne==3.0.2
Pillow>=7.1.2,<7.2.0
cloudinary>=1.26.0,<1.30.0
django-cloudinary-storage>=0.3.0,<0.40.0
coverage==6.2.0
PyYAML>=5.4.1,<5.5.0
uritemplate>=3.0.1.<3.1.0
coreapi>=2.3.3,<2.4.0
black>=22.3.0,<23.0.0
djangorestframework-simplejwt>=4.6.0,<4.7.0
PyJWT>=2.0.1,<2.1.0